+++ react_unity/package.json
@@ -1,34 +1,21 @@
 {
-  "name": "with-create-react-app",
-  "version": "0.1.81",
+  "name": "react_unity",
+  "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@rainbow-me/rainbowkit": "^2.0.2",
-    "@rainbow-me/rainbowkit-siwe-next-auth": "^0.4.0",
-    "@tanstack/react-query": "^5.28.4",
-    "@testing-library/jest-dom": "^6.2.0",
-    "@testing-library/react": "^14.1.2",
-    "@testing-library/user-event": "^14.5.2",
-    "@types/jest": "^29.5.2",
-    "@types/node": "^18.19.3",
-    "@types/react": "^18.2.64",
-    "buffer": "npm:buffer@6.0.3",
-    "ethers": "5.7.2",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-scripts": "5.0.1",
+    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-scripts": "5.0.0",
     "react-unity-webgl": "^8.7.4",
-    "typescript": "5.4.2",
-    "util": "0.12.4",
-    "viem": "^2.8.12",
-    "wagmi": "^2.5.11",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "dev": "react-scripts start",
     "start": "react-scripts start",
     "build": "react-scripts build",
-    "test": "react-scripts test --transformIgnorePatterns \"node_modules/(?!@rainbow-me)/\"",
+    "test": "react-scripts test",
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
