+++ react_unity/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "RainbowKit App",
-  "name": "RainbowKit App Example",
+  "short_name": "React App",
+  "name": "Create React App Sample",
   "icons": [
     {
       "src": "favicon.ico",
