--- react_unity/src/assets/definitions/abi/barcarat.json
@@ -1,296 +1,0 @@
-[
-   {
-      "anonymous":false,
-      "inputs":[
-         {
-            "indexed":true,
-            "internalType":"address",
-            "name":"bettor",
-            "type":"address"
-         },
-         {
-            "indexed":false,
-            "internalType":"uint256",
-            "name":"playerWin",
-            "type":"uint256"
-         },
-         {
-            "indexed":false,
-            "internalType":"uint256",
-            "name":"backerWin",
-            "type":"uint256"
-         },
-         {
-            "indexed":false,
-            "internalType":"uint256",
-            "name":"Tie",
-            "type":"uint256"
-         },
-         {
-            "indexed":false,
-            "internalType":"uint256",
-            "name":"playerPair",
-            "type":"uint256"
-         },
-         {
-            "indexed":false,
-            "internalType":"uint256",
-            "name":"bankerPair",
-            "type":"uint256"
-         },
-         {
-            "indexed":false,
-            "internalType":"uint256",
-            "name":"winAmount",
-            "type":"uint256"
-         },
-         {
-            "indexed":false,
-            "internalType":"enum Baccarat.PairOutcome",
-            "name":"pairOutcome",
-            "type":"uint8"
-         },
-         {
-            "indexed":false,
-            "internalType":"enum Baccarat.GameOutcome",
-            "name":"gameOutcome",
-            "type":"uint8"
-         }
-      ],
-      "name":"BetResult",
-      "type":"event"
-   },
-   {
-      "inputs":[
-         
-      ],
-      "name":"INITIAL_POINTS",
-      "outputs":[
-         {
-            "internalType":"uint256",
-            "name":"",
-            "type":"uint256"
-         }
-      ],
-      "stateMutability":"view",
-      "type":"function"
-   },
-   {
-      "inputs":[
-         
-      ],
-      "name":"addPlayer",
-      "outputs":[
-         
-      ],
-      "stateMutability":"nonpayable",
-      "type":"function"
-   },
-   {
-      "inputs":[
-         
-      ],
-      "name":"bankerPairRate",
-      "outputs":[
-         {
-            "internalType":"uint256",
-            "name":"",
-            "type":"uint256"
-         }
-      ],
-      "stateMutability":"view",
-      "type":"function"
-   },
-   {
-      "inputs":[
-         
-      ],
-      "name":"bankerWinRate",
-      "outputs":[
-         {
-            "internalType":"uint256",
-            "name":"",
-            "type":"uint256"
-         }
-      ],
-      "stateMutability":"view",
-      "type":"function"
-   },
-   {
-      "inputs":[
-         {
-            "internalType":"uint256",
-            "name":"playerWin",
-            "type":"uint256"
-         },
-         {
-            "internalType":"uint256",
-            "name":"backerWin",
-            "type":"uint256"
-         },
-         {
-            "internalType":"uint256",
-            "name":"Tie",
-            "type":"uint256"
-         },
-         {
-            "internalType":"uint256",
-            "name":"playerPair",
-            "type":"uint256"
-         },
-         {
-            "internalType":"uint256",
-            "name":"bankerPair",
-            "type":"uint256"
-         }
-      ],
-      "name":"bet",
-      "outputs":[
-         {
-            "internalType":"uint256",
-            "name":"winAmount",
-            "type":"uint256"
-         },
-         {
-            "internalType":"enum Baccarat.PairOutcome",
-            "name":"pairOutcome",
-            "type":"uint8"
-         },
-         {
-            "internalType":"enum Baccarat.GameOutcome",
-            "name":"gameOutcome",
-            "type":"uint8"
-         }
-      ],
-      "stateMutability":"nonpayable",
-      "type":"function"
-   },
-   {
-      "inputs":[
-         
-      ],
-      "name":"betResult",
-      "outputs":[
-         {
-            "internalType":"enum Baccarat.GameOutcome",
-            "name":"",
-            "type":"uint8"
-         }
-      ],
-      "stateMutability":"view",
-      "type":"function"
-   },
-   {
-      "inputs":[
-         
-      ],
-      "name":"pairResult",
-      "outputs":[
-         {
-            "internalType":"enum Baccarat.PairOutcome",
-            "name":"",
-            "type":"uint8"
-         }
-      ],
-      "stateMutability":"view",
-      "type":"function"
-   },
-   {
-      "inputs":[
-         
-      ],
-      "name":"playerPairRate",
-      "outputs":[
-         {
-            "internalType":"uint256",
-            "name":"",
-            "type":"uint256"
-         }
-      ],
-      "stateMutability":"view",
-      "type":"function"
-   },
-   {
-      "inputs":[
-         
-      ],
-      "name":"playerWinRate",
-      "outputs":[
-         {
-            "internalType":"uint256",
-            "name":"",
-            "type":"uint256"
-         }
-      ],
-      "stateMutability":"view",
-      "type":"function"
-   },
-   {
-      "inputs":[
-         {
-            "internalType":"address",
-            "name":"",
-            "type":"address"
-         }
-      ],
-      "name":"players",
-      "outputs":[
-         {
-            "internalType":"uint256",
-            "name":"points",
-            "type":"uint256"
-         }
-      ],
-      "stateMutability":"view",
-      "type":"function"
-   },
-   {
-      "inputs":[
-         
-      ],
-      "name":"rateDecimals",
-      "outputs":[
-         {
-            "internalType":"uint256",
-            "name":"",
-            "type":"uint256"
-         }
-      ],
-      "stateMutability":"view",
-      "type":"function"
-   },
-   {
-      "inputs":[
-         
-      ],
-      "name":"tieRate",
-      "outputs":[
-         {
-            "internalType":"uint256",
-            "name":"",
-            "type":"uint256"
-         }
-      ],
-      "stateMutability":"view",
-      "type":"function"
-   },
-   {
-      "inputs":[
-         {
-            "internalType":"uint256",
-            "name":"",
-            "type":"uint256"
-         }
-      ],
-      "name":"topThreePlayers",
-      "outputs":[
-         {
-            "internalType":"address",
-            "name":"",
-            "type":"address"
-         }
-      ],
-      "stateMutability":"view",
-      "type":"function"
-   }
-]
