--- react_unity/src/assets/definitions/constants/bacarat.js
@@ -1,5 +1,0 @@
-const bacaratAddress = "0x5102F05f5971975709f2cFEe0B8CbDAd7063Ae16";
-
-export {
-  bacaratAddress
-}
