--- react_unity/src/polyfills.ts
@@ -1,7 +1,0 @@
-import { Buffer } from 'buffer';
-
-window.global = window.global ?? window;
-window.Buffer = window.Buffer ?? Buffer;
-window.process = window.process ?? { env: {} }; // Minimal process polyfill
-
-export {};
